# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gis Backend API",
        "description": "Dokumentasi API untuk platform Gis Backend",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/exam/questions": {
            "get": {
                "tags": [
                    "Question"
                ],
                "summary": "Ambil semua soal (opsional filter level)",
                "operationId": "6b948fb7b3685f8f45941d7a9e1f01e9",
                "parameters": [
                    {
                        "name": "level",
                        "in": "query",
                        "description": "Level soal (SD atau SMP)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "SD",
                                "SMP"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar soal berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Question"
                ],
                "summary": "Buat soal baru beserta 4 jawaban",
                "operationId": "eb4b4ac219a9991ba3cfdf874fe2820f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "level",
                                    "type",
                                    "answers"
                                ],
                                "properties": {
                                    "level": {
                                        "type": "string",
                                        "enum": [
                                            "SD",
                                            "SMP"
                                        ]
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "text",
                                            "image"
                                        ]
                                    },
                                    "question_text": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "question_img": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "answers": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "type"
                                            ],
                                            "properties": {
                                                "type": {
                                                    "type": "string",
                                                    "enum": [
                                                        "text",
                                                        "image"
                                                    ]
                                                },
                                                "answer_text": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "answer_img": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "is_correct": {
                                                    "type": "boolean"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "maxItems": 4,
                                        "minItems": 4
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Soal berhasil disimpan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/exam/questions/{id}": {
            "get": {
                "tags": [
                    "Question"
                ],
                "summary": "Ambil detail soal berdasarkan ID",
                "operationId": "1df3e3e99eee6c067abbfefdb7bebadf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID soal",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail soal berhasil diambil"
                    },
                    "404": {
                        "description": "Soal tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Question"
                ],
                "summary": "Perbarui soal dan jawabannya",
                "operationId": "d6bd32bedabb75d8f79e7833178c84ef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID soal",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "level",
                                    "type",
                                    "answers"
                                ],
                                "properties": {
                                    "level": {
                                        "type": "string",
                                        "enum": [
                                            "SD",
                                            "SMP"
                                        ],
                                        "example": "SD"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "text",
                                            "image"
                                        ],
                                        "example": "text"
                                    },
                                    "question_text": {
                                        "type": "string",
                                        "example": "Apa ibukota Indonesia?",
                                        "nullable": true
                                    },
                                    "question_img": {
                                        "type": "string",
                                        "example": "http://example.com/image.png",
                                        "nullable": true
                                    },
                                    "answers": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "type",
                                                "is_correct"
                                            ],
                                            "properties": {
                                                "type": {
                                                    "type": "string",
                                                    "enum": [
                                                        "text",
                                                        "image"
                                                    ],
                                                    "example": "text"
                                                },
                                                "answer_text": {
                                                    "type": "string",
                                                    "example": "Jakarta",
                                                    "nullable": true
                                                },
                                                "answer_img": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "is_correct": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "maxItems": 4,
                                        "minItems": 4
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Soal berhasil diperbarui"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Question"
                ],
                "summary": "Hapus soal berdasarkan ID",
                "operationId": "0afb391331551b46a379aa812a0b48f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID soal",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Soal berhasil dihapus"
                    },
                    "404": {
                        "description": "Soal tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}